# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pocassist Api",
        "title": "Pocassist Api",
        "contact": {},
        "version": "v0.4.0"
    },
    "paths": {
...
        "/api/v1/product/": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "product list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagesize",
                        "name": "pagesize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "新增",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "product add",
                "parameters": [
                    {
                        "description": "plugin",
                        "name": "plugin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rule.Plugin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/scan/list": {
            "post": {
                "description": "传文件：url列表",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan"
                ],
                "summary": "scan list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "扫描类型：multi / all",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "name": "vul_list",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "备注",
                        "name": "remarks",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "target",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/scan/raw": {
            "post": {
                "description": "传文件：请求报文",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "tags": [
                    "Scan"
                ],
                "summary": "scan raw",
                "parameters": [
                    {
                        "type": "string",
                        "description": "扫描类型：multi / all",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "name": "vul_list",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "备注",
                        "name": "remarks",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "target",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/scan/url": {
            "post": {
                "description": "扫描单个url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scan"
                ],
                "summary": "scan url",
                "parameters": [
                    {
                        "description": "扫描参数",
                        "name": "scan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/scan.scanSerializer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/task/": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "task list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Pagesize",
                        "name": "pagesize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/task/{id}/": {
            "delete": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "删除",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "task delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/msg.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "msg.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "rule.Detail": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "links": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "rule.Plugin": {
            "type": "object",
            "properties": {
                "affects": {
                    "description": "影响类型  dir/server/param/url/content",
                    "type": "string"
                },
                "enable": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "jsonPoc": {
                    "description": "json规则",
                    "$ref": "#/definitions/rule.Poc"
                },
                "vulId": {
                    "description": "漏洞编号",
                    "type": "string"
                }
            }
        },
        "rule.Poc": {
            "type": "object",
            "properties": {
                "detail": {
                    "$ref": "#/definitions/rule.Detail"
                },
                "groups": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/rule.Rule"
                        }
                    }
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rule.Rule"
                    }
                },
                "set": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "rule.Rule": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "expression": {
                    "type": "string"
                },
                "follow_redirects": {
                    "type": "boolean"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                }
            }
        },
        "scan.scanSerializer": {
            "type": "object",
            "required": [
                "target",
                "type"
            ],
            "properties": {
                "remarks": {
                    "type": "string"
                },
                "target": {
                    "description": "单个url",
                    "type": "string"
                },
                "type": {
                    "description": "multi or all",
                    "type": "string"
                },
                "vul_list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}